requirements.txt
pycryptodome>=3.15.0  # For AES encryption/decryption
